+++ crates/codeowners/src/codeowners_hash.rs
@@ -4,14 +4,16 @@ use std::collections::BTreeMap;
 
 hash_content!(
     pub struct CodeownersHash<'cfg> {
-        projects: BTreeMap<&'cfg str, &'cfg OwnersConfig>,
-        workspace: &'cfg CodeownersConfig,
+        pub file_exists: bool,
+        pub projects: BTreeMap<&'cfg str, &'cfg OwnersConfig>,
+        pub workspace: &'cfg CodeownersConfig,
     }
 );
 
 impl<'cfg> CodeownersHash<'cfg> {
     pub fn new(workspace: &CodeownersConfig) -> CodeownersHash {
         CodeownersHash {
+            file_exists: false,
             projects: BTreeMap::new(),
             workspace,
         }